Fix all linter issues and remove wasm-runtime replace directive
- Fix fieldalignment for 16 structs across api, runtime, engine, transport
- Fix noctx: use context-aware net.ListenConfig/Dialer in tests
- Fix revive naming: rename stuttering types (FuncsHost->Host, PollHost->Host,
  StdioHost->Host, HTTPRequestPendingOp->RequestPendingOp, WASMFunctionConfig->FunctionConfig)
- Fix revive: reorder context.Context as first parameter
- Fix revive: consistent receiver naming in MockFS
- Fix unconvert: remove unnecessary type conversions
- Remove replace directive, resolve wasm-runtime from GitHub

+++ runtime/lua/engine/payload/convert.go
@@ -21,7 +21,7 @@ type fieldInfo struct {
 
 var structFieldCache sync.Map // map[reflect.Type][]fieldInfo
 
-const maxLuaInteger = uint64(^uint64(0) >> 1)
+const maxLuaInteger = ^uint64(0) >> 1
 
 // getStructFields returns cached field info for a given struct type.
 func getStructFields(rt reflect.Type) []fieldInfo {
